Move duplicate code into new module
